requirements.txt
streamlit==1.26.0
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
oracledb==1.4.0
pandas==2.0.3
openpyxl==3.1.2
python-dateutil==2.8.2